squash: finish switching settheme over to being a do-task
