+++ tutor/templates/hooks/lms/settheme
@@ -1,5 +1,5 @@
 theme=""
-domains=()
+domains=""
 usage="usage: settheme THEME (--domain DOMAIN ...)+"
 while [ $# -gt 0 ]; do
 	case "$1" in
@@ -9,7 +9,7 @@ while [ $# -gt 0 ]; do
 				echo "$usage"
 				exit 1
 			fi
-			domains=($domains $2)
+			domains="$domains $2"
 			shift 2
 			;;
 		-*|--*)
@@ -29,34 +29,37 @@ while [ $# -gt 0 ]; do
 	esac
 done
 if [ -z "$theme" ]; then
-	echo "error: must specify both USERNAME and EMAIL"
+	echo "error: must specify THEME"
 	echo "$usage"
 	exit 1
 fi
-{% raw %}
-if ((${#domains[@]})); then
-{% endraw %}
+if [ -z "$domains" ]; then
 	echo "No domains supplied; nothing to do."
 	echo "$usage"
 	exit 0
-do
+fi
 code=""
-for domain in ${domains[@]}; do
+for domain in $domains; do
 {% raw %}
 	if [ ${#domain} -gt 50 ]; then
 {% endraw %}
 		echo "Warning:"
 		echo "Assigning a theme to a site with a long (> 50 characters) domain name."
         echo "The displayed site name will be truncated to 50 characters."
-	end
-	code="${code}\nprint('Assigning theme ${theme} to ${domain}...')"
-	code="${code}\nsite, _ = Site.objects.get_or_create(domain='${domain}')"
-	code="${code}\nif not site.name:"
-	code="${code}\n    name_max_length = Site._meta.get_field('name').max_length"
-	code="${code}\n    name = '${domain}'[:name_max_length]"
-	code="${code}\n    site.name = name"
-	code="${code}\n    site.save()"
-	code="${code}\nsite.themes.all().delete()"
-	code="${code}\nsite.themes.create(theme_dir_name='${theme}')"
+	fi
+	code="\
+from django.contrib.sites.models import Site
+site, _ = Site.objects.get_or_create(domain='${domain}')
+if not site.name:
+    name_max_length = Site._meta.get_field('name').max_length
+    name = '${domain}'[:name_max_length]
+    site.name = name
+    site.save()
+site.themes.all().delete()
+site.themes.create(theme_dir_name='${theme}')"
+	echo
+	echo "Assigning theme ${theme} to ${domain}:"
+	echo "$code"
+	echo
+	./manage.py lms shell -c "$code"
 done
-./manage.py lms shell -c "$code"
